alpineQuest/requirements.txt
Appium-Python-Client==4.3.0
selenium==4.27.1
pytest==8.3.3
